+++ src/server/server.ts
@@ -72,9 +72,13 @@ import {
     masterSeqNoResolvers,
 } from "./graphql/chain-order"
 import { blockBocResolvers, overrideBlockBocFilter } from "./graphql/block-boc"
-import { createAccountProvider } from "./data/account-provider"
 import { createBocProvider } from "./data/boc-provider"
 import { lastKeyBlockResolvers } from "./graphql/last-key-block"
+import {
+    IAccountProvider,
+    LiteServerProvider,
+    createAccountProvider,
+} from "./data/account-provider"
 
 type QServerOptions = {
     config: QConfig
@@ -341,8 +345,30 @@ export default class TONQServer {
         )
         this.endPoints = []
         this.app = express()
+        this.app.use(express.json({ limit: "50mb" }))
         this.server = http.createServer(this.app)
         const providers = new DataProviderFactory(this.config, this.logs)
+
+        let accountProvider: IAccountProvider | undefined
+        if (options.config.requests.mode == "tcpadnl") {
+            if (!this.liteclient) {
+                throw new Error("liteclient must be set")
+            }
+
+            accountProvider = new LiteServerProvider(
+                this.logs,
+                this.client.boc,
+                this.liteclient,
+            )
+        } else {
+            const provider = createAccountProvider(
+                this.logs,
+                this.config.accountProvider,
+            )
+
+            accountProvider = provider
+        }
+
         this.data =
             options.data ||
             new QBlockchainData({
@@ -355,10 +381,7 @@ export default class TONQServer {
                     "slow",
                 ),
                 blockBocProvider: createBocProvider(this.config.blockBocs),
-                accountProvider: createAccountProvider(
-                    this.logs,
-                    this.config.accountProvider,
-                ),
+                accountProvider: accountProvider,
                 isTests: false,
                 subscriptionsMode: this.config.subscriptionsMode,
                 filterConfig: this.config.queries.filter,