+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.66.0] - 2023-10-18
+
+### New
+
+- Added fields: `BlockchainMessage { src_code_hash  dst_code_hash }`, `BlockchainTransaction { code_hash }`,
+  `Message { src_code_hash dst_code_hash } `, `Transaction { code_hash }`.
+- Added arguments: `blockchain.messages(src_code_hash dst_code_hash)`, `blockchain.transactions(code_hash)`.
+- In `TCP_ADNL` request mode Q-Server uses TON LiteServer API as an account provider.
+
+### Fixed
+
+- Account provider JSON RPC terminated a process in case of request timeout.
+- GraphQL queries with body more than 100KB was failed with `PayloadTooLargeError: request entity too large`.
+  The size of the body is increased up to 50MB.
+
 ## [0.65.4] - 2023-09-29
 
 ### New